move internal/logs to internal/core
