+++ drivers/web/client.go
@@ -10,7 +10,6 @@ import (
 	"github.com/gopherjs/gopherjs/js"
 	"github.com/murlokswarm/app"
 	"github.com/murlokswarm/app/internal/core"
-	"github.com/murlokswarm/app/internal/logs"
 )
 
 var (
@@ -18,8 +17,8 @@ var (
 )
 
 func init() {
-	logger := logs.ToWriter(os.Stderr)
-	logger = logs.WithPrompt(logger)
+	logger := core.ToWriter(os.Stderr)
+	logger = core.WithPrompt(logger)
 	app.Logger = logger
 }
