+++ internal/core/log_test.go
@@ -1,4 +1,4 @@
-package logs
+package core
 
 import (
 	"bytes"
